save position of learning objects in the ui with two way binding #97

+++ AuthoringTool/PresentationLogic/ILearningObjectViewModel.cs
@@ -4,6 +4,6 @@ public interface ILearningObjectViewModel
 {
     string Name { get; set; }
     string Description { get; set; }
-    int PositionX { get; set; }
-    int PositionY { get; set; }
+    double PositionX { get; set; }
+    double PositionY { get; set; }
 }
