+++ AuthoringTool/PresentationLogic/LearningSpace/LearningSpaceViewModel.cs
@@ -25,6 +25,6 @@ public class LearningSpaceViewModel : ILearningObjectViewModel
     public string Authors { get; set; }
     public string Description { get; set; }
     public string Goals { get; set; }
-    public int PositionX { get; set; }
-    public int PositionY { get; set; }
+    public double PositionX { get; set; }
+    public double PositionY { get; set; }
 }
